# docs/swagger.yaml
basePath: /
definitions:
  apiUserModel.SignUpUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      status:
        type: integer
      type:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:50052
info:
  contact: {}
  description: This is a User API microservice
  title: User API
  version: "1.0"
paths:
  /signup:
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apiUserModel.SignUpUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create new user
      tags:
      - user
  /user:
    get:
      consumes:
      - application/json
      description: Get user by name or id
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"